Funciona mais ou menos, mas às vezes dá leak

+++ proj_24-25-p1_base/common/constants.h
@@ -12,5 +12,5 @@
 #define MAX_UNSUBSCRIBE_RESPONSE_SIZE 3 //tamanho max da resposta da mensagem de subscrição (op_code + char result + '\0)
 #define MAX_REQUEST_SIZE 42 //tamanho max para qualquer request
 #define MAX_KEY_SIZE 41 //tamanho max de um key
-#define MAX_DISCONECT_MESSAGE_SIZE 1 //tamanho max da messagem de desconexao (op_code)
+#define MAX_DISCONECT_MESSAGE_SIZE ((size_t) 1) //tamanho max da messagem de desconexao (op_code)
 #define MAX_DISCONECT_RESPONSE_SIZE 3 //tamanho max da messagem de desconexao (op_code + char result + '\0')
